Scheduler fixes (#67)
* Scheduler Updates
* Create a base scheduler that the dispatch scheduler inherits from
* Reset the `current_observation` when going to the `housekeeping` state

* Small fix to reset observation
