+++ pocs/state/states/default/parking.py
@@ -5,9 +5,6 @@ def on_enter(event_data):
         pocs.say("I'm takin' it on home and then parking.")
         pocs.observatory.mount.home_and_park()
 
-        # while not pocs.observatory.mount.is_parked:
-        #     pocs.sleep()
-
         pocs.next_state = 'parked'
 
     except Exception as e:
