+++ pocs/tests/test_base_scheduler.py
@@ -1,0 +1,204 @@
+import pytest
+import yaml
+
+from astropy import units as u
+from astropy.coordinates import EarthLocation
+
+from astroplan import Observer
+
+from pocs.scheduler.scheduler import BaseScheduler as Scheduler
+from pocs.utils.config import load_config
+
+from pocs.scheduler.constraint import Duration
+from pocs.scheduler.constraint import MoonAvoidance
+
+config = load_config()
+
+# Simple constraint to maximize duration above a certain altitude
+constraints = [MoonAvoidance(), Duration(30 * u.deg)]
+
+simple_fields_file = config['directories']['targets'] + '/simple.yaml'
+loc = config['location']
+location = EarthLocation(lon=loc['longitude'], lat=loc['latitude'], height=loc['elevation'])
+observer = Observer(location=location, name="Test Observer", timezone=loc['timezone'])
+
+
+@pytest.fixture()
+def field_list():
+    return yaml.load("""
+    -
+        name: HD 189733
+        position: 20h00m43.7135s +22d42m39.0645s
+        priority: 100
+    -
+        name: HD 209458
+        position: 22h03m10.7721s +18d53m03.543s
+        priority: 100
+    -
+        name: Tres 3
+        position: 17h52m07.02s +37d32m46.2012s
+        priority: 100
+        exp_set_size: 15
+        min_nexp: 240
+    -
+        name: M5
+        position: 15h18m33.2201s +02d04m51.7008s
+        priority: 50
+    -
+        name: KIC 8462852
+        position: 20h06m15.4536s +44d27m24.75s
+        priority: 50
+        exp_time: 60
+        exp_set_size: 15
+        min_nexp: 45
+    -
+        name: Wasp 33
+        position: 02h26m51.0582s +37d33m01.733s
+        priority: 100
+    -
+        name: M42
+        position: 05h35m17.2992s -05d23m27.996s
+        priority: 25
+        exp_time: 240
+    -
+        name: M44
+        position: 08h40m24s +19d40m00.12s
+        priority: 50
+    """)
+
+
+@pytest.fixture
+def scheduler(field_list):
+    return Scheduler(observer, fields_list=field_list, constraints=constraints)
+
+
+def test_scheduler_load_no_params():
+    with pytest.raises(TypeError):
+        Scheduler()
+
+
+def test_no_observer():
+    with pytest.raises(TypeError):
+        Scheduler(fields_file=simple_fields_file)
+
+
+def test_bad_observer():
+    with pytest.raises(TypeError):
+        Scheduler(fields_file=simple_fields_file, constraints=constraints)
+
+
+def test_loading_target_file():
+    scheduler = Scheduler(observer, fields_file=simple_fields_file, constraints=constraints)
+    assert scheduler.observations is not None
+
+
+def test_loading_target_file_via_property():
+    scheduler = Scheduler(observer, fields_file=simple_fields_file, constraints=constraints)
+    scheduler._observations = dict()
+    assert scheduler.observations is not None
+
+
+def test_with_location(scheduler):
+    assert isinstance(scheduler, Scheduler)
+
+
+def test_loading_bad_target_file():
+    with pytest.raises(FileNotFoundError):
+        Scheduler(observer, fields_file='/var/path/foo.bar')
+
+
+def test_new_fields_file(scheduler):
+    scheduler.fields_file = simple_fields_file
+    assert scheduler.observations is not None
+
+
+def test_new_fields_list(scheduler):
+    assert len(scheduler.observations.keys()) > 2
+    scheduler.fields_list = [
+        {'name': 'Wasp 33',
+         'position': '02h26m51.0582s +37d33m01.733s',
+         'priority': '100',
+         },
+        {'name': 'Wasp 37',
+         'position': '02h26m51.0582s +37d33m01.733s',
+         'priority': '50',
+         },
+    ]
+    assert scheduler.observations is not None
+    assert len(scheduler.observations.keys()) == 2
+
+
+def test_scheduler_add_field(scheduler):
+    orig_length = len(scheduler.observations)
+
+    scheduler.add_observation({
+        'name': 'Degree Field',
+        'position': '12h30m01s +08d08m08s',
+    })
+
+    assert len(scheduler.observations) == orig_length + 1
+
+
+def test_scheduler_add_bad_field(scheduler):
+
+    orig_length = len(scheduler.observations)
+    scheduler.add_observation({
+        'name': 'Duplicate Field',
+        'position': '12h30m01s +08d08m08s',
+        'exp_time': -10
+    })
+    assert orig_length == len(scheduler.observations)
+
+
+def test_scheduler_add_duplicate_field(scheduler):
+
+    scheduler.add_observation({
+        'name': 'Duplicate Field',
+        'position': '12h30m01s +08d08m08s',
+    })
+
+    with pytest.raises(AssertionError):
+        scheduler.add_observation({
+            'name': 'Duplicate Field',
+            'position': '12h30m01s +08d08m08s',
+        })
+
+
+def test_scheduler_add_duplicate_field_different_name(scheduler):
+    orig_length = len(scheduler.observations)
+
+    scheduler.add_observation({
+        'name': 'Duplicate Field',
+        'position': '12h30m01s +08d08m08s',
+    })
+
+    scheduler.add_observation({
+        'name': 'Duplicate Field 2',
+        'position': '12h30m01s +08d08m08s',
+    })
+
+    assert len(scheduler.observations) == orig_length + 2
+
+
+def test_scheduler_add_with_exp_time(scheduler):
+    orig_length = len(scheduler.observations)
+
+    scheduler.add_observation({
+        'name': 'Added Field',
+        'position': '12h30m01s +08d08m08s',
+        'exp_time': '60'
+    })
+
+    assert len(scheduler.observations) == orig_length + 1
+    assert scheduler.observations['Added Field'].exp_time == 60 * u.second
+
+
+def test_remove_field(scheduler):
+    orig_keys = list(scheduler.observations.keys())
+
+    # First remove a non-existing field, which should do nothing
+    scheduler.remove_observation('123456789')
+    assert orig_keys == list(scheduler.observations.keys())
+
+    scheduler.remove_observation('HD 189733')
+    assert orig_keys != list(scheduler.observations.keys())
