+++ pocs/tests/test_dispatch_scheduler.py
@@ -73,138 +73,6 @@ def scheduler(field_list):
     return Scheduler(observer, fields_list=field_list, constraints=constraints)
 
 
-def test_scheduler_load_no_params():
-    with pytest.raises(TypeError):
-        Scheduler()
-
-
-def test_no_observer():
-    with pytest.raises(TypeError):
-        Scheduler(fields_file=simple_fields_file)
-
-
-def test_bad_observer():
-    with pytest.raises(TypeError):
-        Scheduler(fields_file=simple_fields_file, constraints=constraints)
-
-
-def test_loading_target_file():
-    scheduler = Scheduler(observer, fields_file=simple_fields_file, constraints=constraints)
-    assert scheduler.observations is not None
-
-
-def test_loading_target_file_via_property():
-    scheduler = Scheduler(observer, fields_file=simple_fields_file, constraints=constraints)
-    scheduler._observations = dict()
-    assert scheduler.observations is not None
-
-
-def test_with_location(scheduler):
-    assert isinstance(scheduler, Scheduler)
-
-
-def test_loading_bad_target_file():
-    with pytest.raises(FileNotFoundError):
-        Scheduler(observer, fields_file='/var/path/foo.bar')
-
-
-def test_new_fields_file(scheduler):
-    scheduler.fields_file = simple_fields_file
-    assert scheduler.observations is not None
-
-
-def test_new_fields_list(scheduler):
-    assert len(scheduler.observations.keys()) > 2
-    scheduler.fields_list = [
-        {'name': 'Wasp 33',
-         'position': '02h26m51.0582s +37d33m01.733s',
-         'priority': '100',
-         },
-        {'name': 'Wasp 37',
-         'position': '02h26m51.0582s +37d33m01.733s',
-         'priority': '50',
-         },
-    ]
-    assert scheduler.observations is not None
-    assert len(scheduler.observations.keys()) == 2
-
-
-def test_scheduler_add_field(scheduler):
-    orig_length = len(scheduler.observations)
-
-    scheduler.add_observation({
-        'name': 'Degree Field',
-        'position': '12h30m01s +08d08m08s',
-    })
-
-    assert len(scheduler.observations) == orig_length + 1
-
-
-def test_scheduler_add_bad_field(scheduler):
-
-    orig_length = len(scheduler.observations)
-    scheduler.add_observation({
-        'name': 'Duplicate Field',
-        'position': '12h30m01s +08d08m08s',
-        'exp_time': -10
-    })
-    assert orig_length == len(scheduler.observations)
-
-
-def test_scheduler_add_duplicate_field(scheduler):
-
-    scheduler.add_observation({
-        'name': 'Duplicate Field',
-        'position': '12h30m01s +08d08m08s',
-    })
-
-    with pytest.raises(AssertionError):
-        scheduler.add_observation({
-            'name': 'Duplicate Field',
-            'position': '12h30m01s +08d08m08s',
-        })
-
-
-def test_scheduler_add_duplicate_field_different_name(scheduler):
-    orig_length = len(scheduler.observations)
-
-    scheduler.add_observation({
-        'name': 'Duplicate Field',
-        'position': '12h30m01s +08d08m08s',
-    })
-
-    scheduler.add_observation({
-        'name': 'Duplicate Field 2',
-        'position': '12h30m01s +08d08m08s',
-    })
-
-    assert len(scheduler.observations) == orig_length + 2
-
-
-def test_scheduler_add_with_exp_time(scheduler):
-    orig_length = len(scheduler.observations)
-
-    scheduler.add_observation({
-        'name': 'Added Field',
-        'position': '12h30m01s +08d08m08s',
-        'exp_time': '60'
-    })
-
-    assert len(scheduler.observations) == orig_length + 1
-    assert scheduler.observations['Added Field'].exp_time == 60 * u.second
-
-
-def test_remove_field(scheduler):
-    orig_keys = list(scheduler.observations.keys())
-
-    # First remove a non-existing field, which should do nothing
-    scheduler.remove_observation('123456789')
-    assert orig_keys == list(scheduler.observations.keys())
-
-    scheduler.remove_observation('HD 189733')
-    assert orig_keys != list(scheduler.observations.keys())
-
-
 def test_get_observation(scheduler):
     time = Time('2016-08-13 10:00:00')
